third_party/WebKit/LayoutTests/fast/dom/HTMLDialogElement/dialog-open.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<dialog id="mydialog">It's my dialog.</dialog>
<script>
description("Tests that dialog is visible after show() is called and not visible after close() is called. bug 90931");

dialog = document.getElementById('mydialog');
computedStyle = window.getComputedStyle(dialog, null);
shouldBe("computedStyle.getPropertyValue('display')", "'none'");

dialog.show();
computedStyle = window.getComputedStyle(dialog, null);
shouldBe("computedStyle.getPropertyValue('display')", "'block'");

dialog.close();
computedStyle = window.getComputedStyle(dialog, null);

shouldBe("computedStyle.getPropertyValue('display')", "'none'");
shouldThrow('dialog.close();', "'InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable.'");
</script>
</body>
</html>
